+++ src/cmdliner.ml
@@ -7,6 +7,7 @@
 module Manpage = Cmdliner_manpage
 module Arg = Cmdliner_arg
 module Term = struct
+  type ('a, 'b) stdlib_result = ('a, 'b) result
 
   include Cmdliner_term
 
@@ -111,7 +112,7 @@ module Term = struct
     ('a, [ term_escape
          | `Exn of exn * Printexc.raw_backtrace
          | `Parse of string
-         | `Std_help of Manpage.format | `Std_version ]) Pervasives.result
+         | `Std_help of Manpage.format | `Std_version ]) stdlib_result
 
   let run ~catch ei cl f = try (f ei cl :> 'a eval_result) with
   | exn when catch ->
@@ -281,9 +282,10 @@ module Term = struct
   | `Ok n -> n
   | r -> exit_status_of_result ?term_err r
 
-  let exit ?term_err r = Pervasives.exit (exit_status_of_result ?term_err r)
+  let stdlib_exit = exit
+  let exit ?term_err r = stdlib_exit (exit_status_of_result ?term_err r)
   let exit_status ?term_err r =
-    Pervasives.exit (exit_status_of_status_result ?term_err r)
+    stdlib_exit (exit_status_of_status_result ?term_err r)
 
 end