+++ CHANGES.md
@@ -1,5 +1,6 @@
 
 
+- Fix 4.08 `Pervasives` deprecation.
 - Fix 4.03 String deprecations.
 - Fix boostrap build in absence of dynlink.
 - Make the `Makefile` bootstrap build reproducible.